Self-managed camera monitoring: remove shell-launched rpicam-vid, auto-restart on TCP client disconnect
- Remove rpicam-vid launch from scripts/start.sh (camera fully Python-owned)
- Add async monitor task in CameraService: polls process.poll() every 1s,
  restarts immediately on exit without incrementing crash counter
- Monitoring auto-starts via _try_start_monitoring() when event loop present
- Remove _find_existing_process(), _pid attr (no longer shared with shell)
- Add _get_pid() helper, make stop_stream/cleanup cancel monitoring first
- 727 tests pass, 98% coverage

+++ src/config_validator.py
@@ -100,7 +100,7 @@ class RetryConfig(BaseModel):
 
 class CircuitBreakerConfig(BaseModel):
     failure_threshold: int = Field(5, ge=1, le=20)
-    recovery_timeout: int = Field(60, ge=5, le=600)
+    recovery_timeout: float = Field(60, ge=5, le=600)
     success_threshold: int = Field(2, ge=1, le=10)
